Initial commit
A simple simulation of Conway's Game of Life using matplotlib. The user cannot set the initial state, instead a board is randomly generated when the programme is started.
The programme allows the user to set the refresh speed, the size of the board, and the probability for each cell in the randomly generated map to be a live cell.