Added significant performance improvements and a nicer board style.
Performance: Changed the way in which the scatterplot of cells is drawn
in each iteration. Before, the plot would be completely cleared and all
the modifications made on the axes was lost, meaning it had to be done
all over again. Now, the plot is cleaned by removing the first child of
the axes which is the PathCollection of the various points. This way,
all the attributes of the axes is kept and only the new points need to
be added.

Interface: The grid colour has been changed to grey. Additionally, a
scale factor now exists to allow the size of the cells and the thickness
of both the grid and axes' spines to be updated depending on the board
dimensions provided on launch. The scale factor does not come into play
during window resizing and therefore the appearance does not maintain
composure through window resizing.